Move the remaining data (data for the Nav and Tech section on the main page) from the pageData file into their respective components, and delete the no unused pageData file
