+++ src/app/components/Tech.tsx
@@ -2,7 +2,44 @@
 import { FC } from 'react';
 import { motion } from 'framer-motion';
 import { slideInFastFL } from '../utils/animations';
-import { techData } from '@/app/pageData';
+import { SiTypescript, SiPostgresql } from 'react-icons/si';
+import { TbBrandNextjs } from 'react-icons/tb';
+import { SiExpress } from 'react-icons/si';
+import { RiTailwindCssFill } from 'react-icons/ri';
+import { BiLogoNodejs } from 'react-icons/bi';
+import { SiJest } from 'react-icons/si';
+import { SiPrisma } from 'react-icons/si';
+import { FaReact } from 'react-icons/fa';
+import { SiStorybook } from 'react-icons/si';
+
+const techData = [
+  {
+    name: 'TypeScript',
+    icon: SiTypescript,
+    link: 'https://www.typescriptlang.org/',
+  },
+  {
+    name: 'Storybook',
+    icon: SiStorybook,
+    link: 'https://storybook.js.org/',
+  },
+  { name: 'React', icon: FaReact, link: 'https://react.dev/' },
+  { name: 'Next.js', icon: TbBrandNextjs, link: 'https://nextjs.org/' },
+  {
+    name: 'Tailwind CSS',
+    icon: RiTailwindCssFill,
+    link: 'https://tailwindcss.com/',
+  },
+  { name: 'Jest', icon: SiJest, link: 'https://jestjs.io/' },
+  { name: 'Node.js', icon: BiLogoNodejs, link: 'https://nodejs.org/en/' },
+  { name: 'Prisma', icon: SiPrisma, link: 'https://www.prisma.io/' },
+  {
+    name: 'PostgreSQL',
+    icon: SiPostgresql,
+    link: 'https://www.postgresql.org/',
+  },
+  { name: 'Express', icon: SiExpress, link: 'https://expressjs.com/' },
+];
 
 const Tech: FC = () => {
   const techIconElements = techData.map((tech, idx) => (
